Rename models.queue to models.post_queue to prevent module shadowing
